+++ src/main/java/lapr/project/model/WorldMap.java
@@ -106,7 +106,7 @@ public class WorldMap {
         double energyCost = 0;
         double totalMass = Constants.SCOOTER_TOTAL_WEIGHT + Constants.DEFAULT_COURIER_WEIGHT;
 
-        Map<Address, Float> orderWeightMap = new HashMap<>();
+        Map<Address, Double> orderWeightMap = new HashMap<>();
         for(Order order : orderList) {
             orderWeightMap.put(order.getClient().getAddress(), order.getTotalWeight());
             totalMass += order.getTotalWeight();
